Owner credentials moved into User Identity table
Since most of our authenticators use the UserIdentity table, it makes
sense for us to store a user's credentials (login, salt, password)
there as well. This patch moves the credentials fields into the
identity table directly, since we don't want to put them in claims
(which are search indexed).
